requirements.txt
pika
opencv-python
azure-storage-blob